create channel handler moved to separate file
